# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Wifi Dongle Rental Application \n[http://swagger.io](http://swagger.io)\
    \ or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  version: "1.0.0"
  title: "Dongle Rental Application"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/GreendayEngineering/DongleApp/1.0.0"
...
schemes:
- "https"
- "http"
paths:
...
  /dongle:
    post:
      tags:
      - "dongle"
      summary: "add a new dongle"
      operationId: "addDongle"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "dongle to be added"
        required: true
        schema:
          $ref: "#/definitions/Dongle"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Dongle"
  /dongle/findByStatus:
    get:
      tags:
      - "dongle"
      summary: "Finds Dongles by status"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findPetsByStatus"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "status"
        in: "query"
        description: "Status values that need to be considered for filter"
        required: true
        type: "array"
        items:
          type: "string"
          default: "available"
          enum:
          - "available"
          - "pending"
          - "sold"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Dongle"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "Dongle"
  /dongle/findByTags:
    get:
      tags:
      - "dongle"
      summary: "Finds Dongles by tags"
      description: "Muliple tags can be provided with comma separated strings. Use\\\
        \ \\ tag1, tag2, tag3 for testing."
      operationId: "findDonglesByTags"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "tags"
        in: "query"
        description: "Tags to filter by"
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Dongle"
        400:
          description: "Invalid tag value"
      deprecated: true
      x-swagger-router-controller: "Dongle"
  /dongle/{dongleId}:
    get:
      tags:
      - "dongle"
      summary: "Find dongle by ID"
      description: "Returns a single dongle"
      operationId: "getDongleById"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "dongleId"
        in: "path"
        description: "ID of dongle to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Dongle"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Pet not found"
      security:
      - api_key: []
      x-swagger-router-controller: "Dongle"
    delete:
      tags:
      - "dongle"
      summary: "Deletes a dongle"
      operationId: "deleteDongle"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "dongleId"
        in: "path"
        description: "Dongle id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Dongle not found"
      x-swagger-router-controller: "Dongle"
  /dongle/{dongleId}/uploadImage:
    post:
      tags:
      - "dongle"
      summary: "uploads an image"
      operationId: "uploadFile"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "dongleId"
        in: "path"
        description: "ID of dongle to update"
        required: true
        type: "integer"
        format: "int64"
      - name: "additionalMetadata"
        in: "formData"
        description: "Additional data to pass to server"
        required: false
        type: "string"
      - name: "file"
        in: "formData"
        description: "file to upload"
        required: false
        type: "file"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Dongle"
...
securityDefinitions:
  dongle_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:dongles: "modify dongles in your account"
      read:dongles: "read your dongles"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"
    example:
      name: "name"
      id: 6
...
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
    example:
      name: "name"
      id: 1
  Dongle:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      category:
        $ref: "#/definitions/Category"
      name:
        type: "string"
        example: "Three Dongle"
      photoUrls:
        type: "array"
        xml:
          name: "photoUrl"
          wrapped: true
        items:
          type: "string"
      tags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Tag"
      status:
        type: "string"
        description: "dongle status in the store"
        enum:
        - "available"
        - "pending"
        - "rented"
    xml:
      name: "Dongle"
    example:
      photoUrls:
      - "photoUrls"
      - "photoUrls"
      name: "Three Dongle"
      id: 0
      category:
        name: "name"
        id: 6
      tags:
      - name: "name"
        id: 1
      - name: "name"
        id: 1
      status: "available"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
    example:
      code: 0
      type: "type"
      message: "message"
...